Read excel file and save to database and vice versa.

+++ AdvancedConcept/ReadFile/model/financial.js
@@ -22,14 +22,13 @@ Financial.init({
         
     },
     unit_sold:{
-        type:DataTypes.INTEGER,
+        type:DataTypes.DECIMAL,
         allowNull:false
     },
     sale_price:{
-        type:DataTypes.INTEGER,
+        type:DataTypes.DECIMAL,
         allowNull:false
     }
-
 },
 {
     sequelize,
